+++ mingling_core/src/asset.rs
@@ -16,3 +16,6 @@ pub mod node;
 
 #[doc(hidden)]
 pub mod renderer;
+
+#[doc(hidden)]
+pub mod help;